+++ test/vitalTrackerServiceTest.js
@@ -1,0 +1,35 @@
+describe('vitalTrackerService service', () => {
+    
+    // var vitalTrackerService;
+
+    // beforeEach(angular.mock.module('app.services'));
+
+    // beforeEach(inject((__vitalTrackerService__) => {
+    //     vitalTrackerService = __vitalTrackerService__;
+    // }));
+
+    // it('should exist', () => {
+    //     expect(vitalTrackerService).toBeDefined();
+    // });
+
+    //  describe('.localCalculate()', () => {
+        
+    //     it('.localCalculate() should be defined', () => {
+    //         expect(vitalTrackerService.localCalculate).toBeDefined();
+    //     })
+        
+    //     it('has calculated calorie intake', () => {
+    //         var data = {
+    //             age: 19,
+    //             gender: 'male',
+    //             height: {
+    //                 feet: 5,
+    //                 inches: 10
+    //             },
+    //             weight: 183,
+    //             exerciseLevel: 1.550 //Daily
+    //         };
+    //         expect(vitalTrackerService.localCalculate(data)).toBeGreaterThan(25.00);
+    //     });
+    // });
+})